#64-1 [Init] Payment POST 요청에 필요한 DTO 구조 생성
