+++ src/main/java/com/project/storemanager_api/domain/pay/dto/request/CreatePayRequestDto.java
@@ -4,8 +4,6 @@ import lombok.*;
 
 import java.util.List;
 
-import static com.project.storemanager_api.domain.pay.entity.Payment.PaymentType;
-
 @Getter
 @ToString
 @EqualsAndHashCode
@@ -18,18 +16,6 @@ public class CreatePayRequestDto {
     private String placeName;
     private Integer totalAmount;
     private Integer discountAmount;
-    private List<PayRequestDto> payList;
-
-
-    @Getter
-    @ToString
-    @EqualsAndHashCode
-    @NoArgsConstructor
-    @AllArgsConstructor
-    @Builder
-    private static class PayRequestDto {
-        private Integer paidMoney;
-        private PaymentType paymentType;
-    }
+    private List<CreatePaymentDetailDto> payList;
 
 }
