+++ src/main/java/com/project/storemanager_api/domain/pay/entity/Payment.java
@@ -16,15 +16,8 @@ public class Payment {
 
     private Long paymentId;
     private Long orderId; // 주문과 연결 1:1관계, NOT NULL
-    private PaymentType paymentType;
     private Integer paymentAmount; // 해당 결제 방식의 결제 금액 (총 금액), NOT NULL
     private Integer discountAmount; // 할인 결제인 경우
     private LocalDateTime createdAt;
 
-    public enum PaymentType {
-        CASH // 현금
-        , CARD // 카드
-        , DIVIDE // 분할결제
-        , MIX // 현금 + 카드
-    }
 }
